+++ CHANGELOG.md
@@ -8,6 +8,7 @@ BREAKING CHANGES
 - [core] Multisig wallets
 - [core] Sub coin reserve and supply on slash
 - [core] Change unbond time for testnet to 720 blocks
+- [core] Limit candidates count to validatorsLimit*3 at given block
 
 IMPROVEMENT